+++ sportsclub/sportsclub-hibernate-dao/src/main/java/org/jboss/snowdrop/samples/sportsclub/dao/hibernate/HibernateReservationRepository.java
@@ -15,7 +15,7 @@ import java.util.Date;
 
 /**
  */
-public class HibernateReservationRepository extends HibernateRepository<Reservation, Integer> implements ReservationRepository
+public class HibernateReservationRepository extends HibernateRepository<Reservation, Long> implements ReservationRepository
 {
 
    public HibernateReservationRepository()
